Refactor product and order status APIs to service modules
Moved all product-related API functions from src/api/api.js to a new src/api/service/ProductService.js file and order status APIs to PaymentService.js. This improves code organization and separation of concerns, making API services more modular and maintainable.
