+++ src/pages/layouts/AdminLayouts.jsx
@@ -10,6 +10,7 @@ import { Toaster } from "react-hot-toast";
 import { AdminAPIProvider } from "@/context/AdminAPIContext";
 import { AdminModalProvider } from "@/context/AdminModalContext";
 import AdminModalTemplate from "@/components/Admin/AdminModalTemplate";
+import { APIOrderProvider } from "@/context/APIOrderContext";
 
 const AdminLayouts = ({ children }) => {
     // Kiểm tra breakpoint là màn hình desktop
@@ -33,37 +34,39 @@ const AdminLayouts = ({ children }) => {
         <AdminModalProvider>
             <ThemeProvider>
                 <AdminAPIProvider>
-                    <div className="min-h-screen bg-slate-100 transition-colors dark:bg-slate-950">
-                        {/* Nền đen */}
-                        <div
-                            className={cn(
-                                `pointer-events-none fixed inset-0 -z-10 bg-black opacity-0 transition-opacity`,
-                                !collapsed && "max-md:pointer-events-auto max-md:z-50 max-md:opacity-30",
-                            )}
-                        />
-
-                        {/* 1. Thanh Sidebar */}
-                        <AppSidebar
-                            ref={sidebarRef}
-                            collapsed={collapsed}
-                        />
-
-                        {/* 2. Phần Navbar và Content */}
-                        <div className={cn(`transition-[margin] duration-300`, collapsed ? "md:ml-[70px]" : "md:ml-[240px]")}>
-                            {/* 2.1: Header Navbar */}
-                            <NavbarAdmin
-                                collapsed={collapsed}
-                                setCollapsed={setCollapsed}
+                    <APIOrderProvider>
+                        <div className="min-h-screen bg-slate-100 transition-colors dark:bg-slate-950">
+                            {/* Nền đen */}
+                            <div
+                                className={cn(
+                                    `pointer-events-none fixed inset-0 -z-10 bg-black opacity-0 transition-opacity`,
+                                    !collapsed && "max-md:pointer-events-auto max-md:z-50 max-md:opacity-30",
+                                )}
                             />
 
-                            {/* 2.2: Phần nội dung */}
-                            <div className="h-[calc(100vh-60px)] overflow-x-hidden overflow-y-auto p-6">
-                                <Outlet />
-                                <Toaster toastOptions={{ duration: 2000 }} /> {/* 5. Thiết lập thông báo Hot Toast Mặc định */}
-                                <AdminModalTemplate />
+                            {/* 1. Thanh Sidebar */}
+                            <AppSidebar
+                                ref={sidebarRef}
+                                collapsed={collapsed}
+                            />
+
+                            {/* 2. Phần Navbar và Content */}
+                            <div className={cn(`transition-[margin] duration-300`, collapsed ? "md:ml-[70px]" : "md:ml-[240px]")}>
+                                {/* 2.1: Header Navbar */}
+                                <NavbarAdmin
+                                    collapsed={collapsed}
+                                    setCollapsed={setCollapsed}
+                                />
+
+                                {/* 2.2: Phần nội dung */}
+                                <div className="h-[calc(100vh-60px)] overflow-x-hidden overflow-y-auto p-6">
+                                    <Outlet />
+                                    <Toaster toastOptions={{ duration: 2000 }} /> {/* 5. Thiết lập thông báo Hot Toast Mặc định */}
+                                    <AdminModalTemplate />
+                                </div>
                             </div>
                         </div>
-                    </div>
+                    </APIOrderProvider>
                 </AdminAPIProvider>
             </ThemeProvider>
         </AdminModalProvider>
